requirements.txt
numpy==1.24.3
openai==0.27.6
pyaml==23.5.9